Lock build/test toolchain and dependencies.
As I look towards automated liniting in Travis, I want to make sure that
tests aren't going to break in unexpected fashions when the upstream
test tooling is updated unexpectedly. Thus we lock these version using
the output of `pip list --format freeze`.

This means that these version will need to be periodically updated to
capture the latest & greatest work, but those updates will happen in a
controlled fashion (a git commit).